resource damage

+++ Assets/Scripts/Player/PlayData.cs
@@ -10,8 +10,8 @@ public class PlayData : MonoBehaviour
     
     public int maxHealth;
     public float moveSpeed;
-    public float axeDamage;
-    public float pickaxeDamage;
+    public int axeDamage;
+    public int pickaxeDamage;
     public float rockHarvestingSpeed;
     public float woodHarvestingSpeed;
     public int resourceCarryLimit;
